requirements.txt
configparser==5.3.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
numpy==1.26.3
pandas==2.2.0
Pillow==9.2.0
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyqt5-tools==5.15.9.3.3
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
qtconsole==5.5.1
QtPy==2.4.1
taichi==1.4.1
